add: full attachment support
supports: photo, video, arbitrary files.
because of the default filesize limit of 50mb, the bot now depends on the local bot API server, which must be launched before launching the bot.

the project now requires yt-dlp library in order to download video files.

all video and "big" files(>1mb) are now temporarily stored in temp_video and temp_doc directories, and *should* be deleted by the bot after they are uploaded to Telegram servers and are safe to delete. Maybe, in the future, temporary documents will be sent as URLs, not saved on the disk.